Some Druid tweaks, enhancements while changing form.
This is a significant tweak to the Druid role, and to comes from
discussions I've had with a few other variant developers about this
role. The gist of it was this - how can we set the Druid role apart from
other roles, so that it's not another 'grab artifact weapon, ascension
kit, and go'? That's a non-trivial task. This is my initial attempt at
it.

The breakdown:
* Renamed #shapechange to #wildshape as this is how it's referred to in
ad&d, and will be called such moving forward
* when a Druid uses wildshape, they gain bonuses to both damage output
and to-hit, and the to-hit bonus scales as the Druid gains experience
levels. At experience level 30, several forms will have a to-hit value
that's on par with a fully kitted out level 30 adventurer swinging a
fully enchanted artifact weapon
* weapon skills have been adjusted - scimitar/saber and quarterstaff are
capped at skilled, all the rest are dropped to basic if they weren't
already. Skill in dart was removed (more incentive to use a sling)
* Enchantment spell skill has been added at basic skill
* normal amount of turns when polymorphed into a monster is 500-1000;
for Druid using #wildshape they enjoy 2500-3000 turns before they revert
back

Perhaps more could be done here to really set Druid apart, but I feel
like this is a good start. We'll see how this is received.