Fix: oversight in rejectcasting() function.
I noticed this while playing with Druid's #wildshape ability - I could
still cast spells in most of the forms I could change into. I'm thinking
to myself 'this can't be right, none of the Druid forms have hands or
even arms sometimes'. And then I had a look in the rejectcasting()
function and saw that only !freehand() was used instead of !u_handsy()
and saw this comment from the NetHack devteam:

    But why isn't lack of free arms (for gesturing) an issue when
    poly'd hero has no limbs?

Yeah. So I tested a couple other scenarios, turns out the player can
still cast spells when poly'd into ANY monster as long as they can
verbalize. I consider this to be a vanilla NetHack bug.

In this commit, I've fixed it by swapping out !freehand() with
!u_handsy(), but added an exception for Druids using #wildshape, as this
adheres to ad&d canon.