Highlight hidden stairs.
From xNetHack commits 24c092b and 9961eb6 (slightly modified).

The player can set an option to highlight stairs that are hidden by
objects on top it. The background is red, just as it is in other
variants that have this feature.

Default behavior is false. Can be enabled in-game via menu, or place
OPTIONS=hilite_hidden_stairs in your config file.