Changes to forging, new object: blacksmith hammer.
This commits reigns in artifact forging (and to a much lesser
degree, forging in general) in the way that there is now a third
'component' needed to forge artifacts - magic.

The changes:
* new object - blacksmith hammer, tool class. This is the only object
that can be used to forge anything. Regular hammers no longer suffice
* the blacksmith hammer can always be found in Goblintown, colocated
with the Blacksmith's Cookbook
* the blacksmith hammer does not randomly generate (meaning it can't be
polypiled for), nor will it ever be found in a bones pile
* this hammer is made of iron, and can be rusted away if the player
isn't careful with it, so it's advised to rustproof it when possible
* when fully identified, players will notice that the blacksmith hammer
takes charges like other various tools (fire/frost horn, can of grease,
etc)
* Charges are required ONLY to forge artifacts, and as a forged artifact
is made, a charge is used up. Regular forged objects do not require nor
use up charges
* Reading a scroll of charging is required to charge the hammer. Reading
a cursed scroll will strip it of charges, an uncursed scroll will grant
2-4 charges, a blessed scroll will grant 4-8 charges
* the blacksmith hammer will spawn with 2-4 charges, and can only hold a
maximum of eight charges (will not blow up if overcharged). Once it's
been recharged twice, it cannot be charged again (similar behavior to
magic markers). This means the maximum number of charges it could ever
have is 20 - since there are less than 20 forged artifacts in the game,
this does not prevent all forged artifacts from being created, so long
as the player is smart (and lucky) with their scrolls of recharging

Forging artifacts requires a bit more work and planning. This is part of
the artifact balancing from previous recent commits.