Forged armor/weapon quality.
With the inclusion of the Blacksmith hammer, I thought it'd be fun to do
some neat things with it given its beatitude.

If the Blacksmith hammer is blessed, when you forge a regular piece or
armor or a weapon, there is a chance for the object to come out as
either 'superior' or 'exceptional'. A superior forged weapon will do an
extra point of damage, and an exceptional forged weapon will do an extra
two points of damage, and gets +1 to-hit. Superior and exceptional
pieces of armor grant a +1/+2 AC bonus respectively.

On the flipside, if the Blacksmith hammer is cursed, there's a good
chance you'll forge an 'inferior' piece or armor or weapon. Inferior
weapons get a -2 penalty to damage and to-hit, and inferior armor gets a
-2 AC penalty.

The odds of obtaining these varying quality objects is - 1 in 10 for a
superior forged object, 1 in 100 for exceptional (blessed hammer), and 4
out of 5 for an inferior object (cursed hammer). Using an uncursed
hammer won't make any quality changes one way or the other.

There's a fair amount of unrealized potential here. Such as weapon
shattering odds based on its quality, their cost in shops, and so forth.
I do plan on implementing things like this for the various quality
objects very soon.