Fix: poison/quality of known artifacts not shown
The black magic of doname and xname was making it so poisoned/tainted, fully known artifacts didn't display their poisoned/tainted status.

This was also the case for inferior/superior/exceptional status, but for a superset of reasons that necessitated removing the display of these qualities from xname entirely (which I think is consistent with the usage of xname anyway? I don't really know what xname does).

They should both display properly now.