Restructure how blacksmith hammer is made unwishable
This makes the blacksmith hammer inherently unique and unwishable. That way it can be treated like similar objects in the wishing code instead of being special-cased. Not really a big difference in functionality from the way it was done before, but it might make the code a bit easier to understand.

Also, you get a normal hammer instead of nothing if you try to wish for a blacksmith hammer. Yay?