Object quality random spawn/wishing rules.
It was starting to feel weird seeing objects that had a quality bit set,
but that same object wasn't anything that could be produced in a forge.
Things like magical objects, and objects that were made of materials
other than what was allowed when forging. That is addressed in this
commit.

Also tweaked the quality type a bit on random generation.