Spread out wand of wishing placement.
This commit changes the guaranteed wand of wishing from being in one
location, to being spread out in three different places. This is to
address having all wishes appear in just one location, and facilitates
the player gradually earning them. It's also a safeguard of sorts in
case one wand is destroyed or used up by a monster. The changes:

* the castle wand is still present, but instead of being 0:x, it's now
flagged as recharged, and has a fixed two charges (1:2)
* the second wand is obtained by destroying Vecna. It's a death drop and
not in his inventory, so it's not something he can use - it is generated
after he's destroyed, just like the Eye/Hand of Vecna. This wand is also
flagged as recharged, and the number of charges will be random (1:x).
This wand is also made erodeproof as Vecna will sometimes fly over lava
and is sometimes destroyed over it
* the third and final wand can be found in the crystal chest in the top
level of the Wizard of Yendor's tower where he spawns. This wand is also
flagged as recharged, and has a fixed charge of one (1:1)

So... before, you had a guaranteed four wishes (0:1, blessed recharge to
1:3) and maybe seven (0:3, blessed recharge to 1:3, get lucky with a
wrest). Now, it's still guaranteed four wish, but maybe nine in total if
Vecna's wand spawns 1:3 and you get lucky wresting all the wands. Sounds
fair to me given who you have to defeat to earn all of these prizes.

I also tweaked random wand of wishing spawning. It's as low as it can be
now; the point I took away from wishing I added to wand of death random
spawn chance.