Revamp player monster spawning gear.
This is a relatively minor bug, but it's existed in EvilHack probbaly
since the very early days of this variant.

What has been noticed for quite some time now, is that higher
level player monsters would have unnecessary extra gear, like an extra
suit of body armor, or an extra bag, and so forth. The reason this was
happening, is because in mplayer.c, this is where player monster gear is
original built. Then in EvilHack, since it became possible for player
monsters to spawn in the regular dungeon and not just on Astral Plane,
has its own gear routine in makemon.c, and those routines would overlap
for the higher level player monsters.

There's no good reason for this. In this commit, I've stripped most of
the gear selection routine out of mplayer.c and put a small bit of it in
makemon.c, and while doing this I made some tweaks. Player monsters
shouldn't spawn with extra bags anymore. There's also a chance the
higher level player monsters will have higher quality weapons.

Might be a bit more fine-tuning that could happen here, but as of now
we're in a much better spot than previously.