Holy water, wands of death, and Draugr.
Potions of holy water will now do more damage to Draugr if quaffed.
Racial zombies (e.g. player monster Draugr) are now also affected if hit
by one.

Awhile ago, I made it so wands of death acted like a potion of full
healing for the undead if they were hit by the beam. That's now true for
Draugr or if the player is poly'd into an undead monster. Just as with
undead monsters, the upper limit is 200 hit points. I added a new flag
for monsters seeing the effects of a wand (M_SEEN_DEATH), so if meet the
critera, and a monsters zaps you with a wand of death, they won't
continue to zap it at you.