Blessed scroll of taming increases tameness of pets.
This pulls in NetHack 3.7 commits 9b4f38e (title commit) as well as
commits 13b8725 and 6df83cb, as they all compliment each other.

Reading an uncursed/blessed scroll of taming, or casting the spell
'charm monster' at skilled or higher in enchantment spells, will
increase the tameness of your pet. The boost for an uncursed scroll is
1, at blessed it's 2. Casting 'charm monster' at skilled or higher
assumes blessed.

While here, have invoking the artifact Bracers of the First Circle act
as if it were a blessed scroll of taming.