Fix: give message when gaining strength via uncursed gain ability
Evilhack introduces a special gainstr() function to handle increased strength gain for giants and such. It handled flagging for messages differently from the general adjattrib() function for attribute changes, which meant it could not tell you if your strength increased when drinking an uncursed potion of gain ability. Make the flag identical to adjattrib()'s.