Allow Druids that are in one of their wildshape forms to manipulate objects and their surroundings.
In building out the Druid role for EvilHack, I've tried to stick to ad&d
rules as closely as I can. One of the abilities that Druids have is
using 'Wildshape', and they assume the form they turn into. NetHack and
it's variants in general already do a pretty good job handling what the
player can and can't do while in various forms. Earlier, I had already
allowed for Druids to be able to cast spells in most of the forms they
can assume, and this is a thing with ad&d Druids also. A point of
contention that was recently brought up was about how almost every shape
a Druid could turn into, they have no hands, which translates into a
wildshaped Druid not being able to do a lot of things, like
opening/closing doors or containers, using a key to unlock/lock a door,
and so on. In ad&d this could work, but in the NetHack space, not as
much as we'd like. I did a little bit of digging, and it turns out there
is precedence for Druids being able to manipulate their surroundings
like they had hands while in a wildshape form. It basically boils down
to the DM's discretion. Here's what I've decided to do:

* Out of the 22 forms a Druid can assume using wildshape, all but two
(snake and giant anaconda) can interact with objects and their
surroundings as if they still had hands. This covers opening and closing
doors and containers, using tools, engraving, throwing things,
untrapping traps, and so forth
* Wielding weapons or wearing a shield is still off-limits, seeing a
pegasus wielding a scimitar is too much
* Snake type monsters were excluded because there's no plausible way to
think they can manipulate anything in any kind of 'hands-on' way

I believe I've covered all the bases, but if I've missed something, it's
easy enough to add a condition for wildshape later on without breaking
saves.