Fix: rehumanizing a Druid while in wildshape via prayer.
In pray.c there's a case for TROUBLE_UNUSEABLE_HANDS, ad when you pray
it checks for two things: 1) if you're wielding a cursed weapon or 2)
you're polymorphed into a form that has no hands. The second is checked
because in every case before Druids and wildshape were invented, having
no hands was extremely limiting. The fix is to simply not consider
Druids using wildshape when this case is checked.