Fix: give proper feedback on why certain objects can't be used while in wildshape.
Recently a suggestion was made to allow Druids in wildshape to do
certain things like #rub a lamp, or apply a pick-axe. The feedback given
when attempting to do these things was 'You can't hold [obj] strongly
enough'. This checks cantwield() and in almost all cases, this stays
true for Druids in wildshape because they're not allowed to wield
weapons. Doing things like rubbing a lamp, applying a digging tool, or
applying a polearm requires that object to be wielded for it to work.
Current behavior is intended, but the feedback a wildshaped Druid
receives should make more sense in that moment.