Remove race flag from Dramborleg.
Reason: Dramborleg is now one of the objects needed to forge Ithilmar.
Before this commit, Dramborg was intelligent, lawful, and attuned to
dwarves. That three-way combo made it impossible to even have it in
inventory if you weren't lawful or a dwarf. Remving one of those three
flags makes it possible for Dramborleg to be used as a forge recipe by
any race/role/alignment combo. You'll still get blasted of course if
non-lawful, but at least it can be handled.