Fix: rotten food showing as 'burnt'.
Commit f869012 basically 'broke' what we had in place earlier to show
oeroded food objects as rotted vs burnt. This commit replaces the
identifier with words that will not interfere with rotten tins of
[something]. I chose 'decayed' for various levels of oeroded, and
'preserved' if the food object has been fooproofed.

The bitfield oeroded is overloaded, as it also handles orotten and
odiluted. This really should be fixed, it would make handling these
types of issues a lot simpler, or make them a non-issue entirely. I may
fix this myself for next version (fixing this proper will break saves).