Fix: limit how often a female Druid can lay eggs while using wildshape.
I was initially going to limit this before public release, but I wanted
to see how this would play out, see if it'd be abused like I thought it
might be. Didn't have to wait long 😁

Female Druids while using wildshape can still attempt to lay an egg, and
they may even be successful, but the odds are not in their favor. Plus
the attempt will take up a turn.