Fix: infinite loop when praying in handsless wildform
When very lucky while praying, your god may attempt to fix *all* your troubles. Handlessness in wildform counted as a trouble but was not fixed, resulting in an infinte loop.

Make it so handless wildforms are not considered a trouble.

Limit the number of times troubles are attempted to be fixed in case a similar bug ever occurs in the future.