Fix: only allow dead tree 'corpses' to form on room/grass/sand tiles.
This was initially reported as having killed a tree blight in a corridor
and then kicking down its dead tree corpse, and having the spot it was
on turn into a room tile. Then it occurred to me that I should test this
out if this had happened on say, the stairs. The stairs glyph would
disappear, and be replaced with a room tile. The stairs were still there
however, and still functioned. Could be worse, but this is not ideal at
all. This commit makes it so that a dead tree corpse is only left
behind on an acceptable tile space.