Fix: digging out traps on no-dig levels.
This commit is very loosely based on NetHack 3.7 commit d5785f2, which
was in response to issue #1254:

'If there's a trap on a no-dig level, the floor beneath it
will always be "too hard to dig into", making it impossible to remove
the trap.

As you can still dig pits in these levels (just not holes), the floor
under the trap itself resisting to become a pit seems inconsistent.'

Doing this because the buried chests at mines' end that contain the
magic key would sometimes become inaccessible due to certain random
traps spawning on the same space. Now the player will be able to dig the
trap away and gain access to the buried chest and its contents.