Mistletoe can affect a Druids power regeneration (Part I).
There's a few things in this commit that all tie together, but the main
bit is how mistletoe helps (or lack thereof hinders) a Druids power
regeneration. I call this 'Part I' because there's more I think that
could be done here.

In ad&d, mistletoe was a required physical component for almost every
spell a Druid would cast. We're not taking it to that level of course,
but it can still play a minor role. Here's how it works.

* a Druid with 9 or more mistletoe in open inventory, their normal spell
power regeneration doubles
* a Druid with 5-8 mistletoe in open inventory, normal spell power
regeneration is roughly 1.3x of normal
* a Druid with one to four mistletoe in open inventory, normal power
regeneration is normal
* a Druid with no mistletoe in open inventory, normal power regeneration
is cut in half

A quirk with NetHack, all of the mistletoe have to be grouped/merged
together. Meaning if you have 5 uncursed mistletoe in one inventory
slot, and also 5 cursed/blessed (or unidentified) mistletoe in another
mistletoe slot, that won't meet the 9 or greater threshold.

Druids already start with two mistletoe. Given the number of live trees
available throughout the game (or a Druid could just make some via
spell), acquiring more mistletoe shouldn't be difficult.

Also in this commit - when kicking a tree for fruit, only one mistletoe
will ever be generated. And when a Druid kicks a tree, they don't get
hurt, and the feedback shows the Druid being nicer to the tree than
other roles doing the same. There's been some discussion about how a
Druid would properly go about obtaining tree fruit, but until a
different method is decided on, this will suffice.

Last bit, the plural of mistletoe is still just mistletoe, at least
that's how I see it. Online search says its 'mistletoes' but that
doesn't feel right. Gonna roll with it for now.