Revamp of crowning and fractured altars.
Previous commit, you could not get crowned on an already fractured
altar, but getting crowned wouldn't fracture said altar. Based on some
player feedback, I'm switching things around.

With this commit, you can get crowned on an altar regardless of whether
it's fractured or not, but if crowned on an intact altar, it will
fracture it.

You may notice on_altar() was used here, this is because of the wizmode
command 'wizcrown' which will crown you whether you're standing on an
altar or not.