The Hand of Vecna has a chance to cause area of effect explosive cold damage.
When attacking a monster via melee, the Hand of Vecna now has a small
chance of an additional explosive area of effect cold damage to its
target. The damage is 2d6 if wielding a weapon, or 4d6 if attacking
bare-handed.

Unlike other artifacts that do this type of additional damage, the Hand
of Vecna will not protect objects in open inventory from cold damage.
This is an evil intelligent artifact, and as such does not care about
your belongings.

I also made a minor fix with Harbinger, added proper feedback when its
explosive AoE acid damage procs.