Health regeneration for Druids while in wildshape.
Normally when a player polymorphs into [insert monster here], their
normal rate of health regeneration is slowed by a fair amount. This of
course affected Druids while using wildshape as well. This commit
changes that, so when a Druid is using wildshape, their health regen
while in wildshape form is faster. Not quite as fast as if in original
form, but pretty close to it. The difference from before to now is
significant and will be easily noticed.