Fix: guard against dead monsters when various artifacts proc their explosive effect.
Recently a player was seeing some dmonsfree errors when killing the same
monster twice with harbinger when its acid area of effect explosive
effect fired off. I wasn't able to recreate the impossible, but this
commit should address the issue.

Also in this commit - the Hand of Vecna's explosive effect won't trigger
if wielding another artifact. It will still occur if wielding a
non-artifact or fighting bare-handed.