Fix: dmonsfree with artifacts that cause explosive damage when thrown.
This issue was still occurring with Harbinger when thrown (it's an
aklys) and it's explosive effect triggered. Easiest/best solution, don't
let the explosive effect proc unless the artifact is actually wielded.
This actually goes in the player's favor - if you were throw harbinger
at melee range and the explosion effect triggers, the player nor their
stuff in open inventory is protected (not wielded).

While here, noticed and fixed a bug with dichotomy. If a monster were
not resistant to both fire and cold damage, only the fire explosive
effect would trigger. Now both fire and cold have an equal chance.