Allow pets that can eat grass, to eat grass when hungry.
If you have a pet, and that pet can eat grass, they will do so if
hungry. The nutrition value gained from doing so is 20 (same as it is
for a player eating grass if polyd into a grass-eating monster).

Non-tame monsters that can eat grass will not do so.

In this current implementation, your pet would have to arbitrarily walk
over a grass tile, or be led to it (ridden, on a leash, etc) before
eating. I haven't figured out how to make a hungry grass-eating pet want
to seek out grass yet, but that's on the to-do.