Add evocation spellbook descriptions to the in-game encyclopedia.
Recently, a player had mentioned how nice it'd be if the newwer spells
added to the game could be looked up in-game via the pokedex
(encyclopedia). I noticed that none of the spells in the game have any
entries at all. Figured it'd be a good idea to remedy that, so... this
commit handles the nine new evocation spells (8 brand new, and 'finger
of death' shifted from attack spell school to evocation). My plan is to
add descriptions for all of the player spells to the pokedex.