Fix: Druid taking blame for pet killing a tree.
Recently, a player playing as a Druid had this happen to them: they have
a pet titan. Pet titan starts attacking other monsters, one of those
attacks, it cast the monster spell 'fire bolt' which causes an
explosion. A live tree was next to the monster being attacked; the
resulting explosion killed the tree. The player took the blame for
killing the tree. Ugh.

This commit fixes that issue, and while here I noticed a couple other
spots could use the same treatment.