Fix: chopping boulder at tree spot.
From NetHack 3.7 commit 6c42180 (slightly modified):

'Reported by k21971:  applying an axe toward a location that contained
both a tree and a boulder (or statue) would use the axe to break the
boulder/statue rather than chop down the tree.

Different code is used to finish the dig/chop than is used to decide
whether the tool is appropriate for its target.'

Was first discovered in EvilHack, as it's not uncommon to kill an
ent-type monster that's carrying around a boulder, and a dead ent leaves
behind a dead tree. Reported the bug to the NetHack devteam as it was
present there as well.