Expert skill in wielded weapons that can be thrown/returned will always return.
Weapons affected by this commit: aklys (to include the forged artifact
Harbinger), Mjollnir, the Hammer of the Gods, and Xiuhcoatl.

If the player has expert or greater skill associated with the weapon
they are throwing, if it can return, it will always return so long as
they are not impaired (confused, stunned, etc).