Fix: dmonsfree with artifacts that cause explosive damage.
This was still occurring, even after all of the previous fixes. Turns
out this issue would present itself if the player was blind and the
explosion effect triggered if the monster being hit died from the
physical act of being hit. In artifact_hit(), many of the special damage
types artifacts deliver return 'realizes_damage' which has a 'can you
see the target' component. The fix was to simply return 'true' after
explode() is called for various artifacts that call it.

As this is the correct and (hopefully) final fix for this, we can revert
commit 47d6282, which prevented the explosion effect to trigger if the
artifact were thrown (and it was happening because players using
Harbinger would use a boulder fort, then throw Harbinger at monsters,
but you can't see what's on the other side of a boulder fort unless you
*blind* yourself and have ESP).