Fix: the Riders/Wizard of Yendor should not be able to open any locked container.
Many moons ago, when I was importing GruntHack's code to allow monsters
to loot containers, there was a bit in that code in mpickstuff() that
gave the Riders and the Wizard of Yendor a free pass so to speak. They
were allowed to open any locked container, without needing any kind of
unlocking tool. Back then, no big deal. Fast-forward to the inclusion of
crystal and magic chests, and one of each being placed right next to
where the Wizard of Yendor spawns...

I'm not seeing any good reason why either should be able to access a
locked container freely like that. That ability has been removed.