Fix: grammar feedback when spells like barkskin or stoneskin wear off.
Scenario: Druid casts barkskin on themselves, then wildshapes into a
lizard or a snake. Spell starts to wear off, feedback is 'Your scales
feels softer' which is all sorts of wrong. Rather than try to juggle the
code in objnam.c to try and determine what to pluralize and what not to,
I'll just reword the feedback so it's plural agnostic. So no matter what
form you're in (skin/scales/bones/hide/etc), the feedback will be
correct.