Fix: impossible "portal on top of portal??" when loading Gehennom level with the portal to Vecna's domain.
I was aware of this bug while working on the previous commit, but this
had a bit lower priority since not having the portal to Vecna's domain
spawn would not make the game unwinnable (Vecna's domain is an optional
branch). The cause of this impossible in this regard is due to the lava
rivers that form in each filler level in Gehennom. Prevent those from
forming, bug goes away. But we want those lava rivers. So... the fix is
to prevent a magic portal from forming on an already occupied space.

The function occupied() checks for pools, lava, furniture, existing
traps, and the vibrating square. Perfect for this scenario. While here,
I made a slight improvement in somexyspace(), and did some minor code
formatting in mklev.c

I made sure magic portals spawn where they're supposed to (all of the
planes, quest entry portal, etc) before committing.