Change access to the treasure room in the Mines' End map 'The Sewers of Waterdeep'.
Accessing the treasure room in the Mines' End map 'The Sewers of
Waterdeep' is no longer achieved by teleporting in. Defeating the Rat
King will create a locked door on the east side of the treasure room,
accessible via the short north/south hallway that is adjacent to the
room.

Instead of the entire map being non-passable (phasing), only the walls
to the treasure room and the Rat King's throne room are non-passable.
The entire map is still non-diggable. Because teleporting in is no
longer the preferred method to access the treasure room, most of the
set teleportation traps have been removed. The one in the treasure room
is still available in case the player somehow winds up inside the room
before killing the Rat King, such as falling down a deep shaft and into
the level. The chance of the Rat King spawning with a ring of teleport
control has been removed.