Druids can pass directly through trees.
Druids in ad&d can walk right through dense vegetation or trees like
they aren't even there, once they reach a certain experience level.
Thought that'd be fun for Druids in EvilHack to be able to do as well.

Once a Druid obtains the rank of Mystic (level 14), they can walk
through trees (dead or alive) as if they weren't there. This is a
permanent ability, no need to have to toggle on or off.

The framework is setup in such a way, that if I wanted to create an
object that could impart this ability, I could do so (or also allow
monsters to use this ability), but for now it's strictly an intrinsic
for player Druids only.

Could be bugs with this, I haven't tested it fully, but did test it
enough that I felt confident enough to commit.