Trap ammo overhaul (xNetHack).
This commit incorporates the trap ammo overhaul patches from xNetHack.
The number of commits made from xNetHack are too numerous to list here
(close to 20), but the first three commits (ab92233, 73b890e, d4a4000)
are a good place to look and see how it started there.

What this does: arrow/bolt/dart/falling rock traps now store a stack of
their respective ammo in a new ammo field created specifically for this.
Bear traps and landmines store a single trap object in their set traps.
For the ammo-based traps, what they spawn with for ammo is what they'll
always have, instead of just making up an amount randomly. So if a dart
trap spawns with 18 darts in it, you walk over it and get hit by one,
then untrap it on the next turn, you'll get 17 darts out of it. If
the trap ammo spawns with a certain material, that should stay
consistent for the entire stack. Spear traps were wrapped up into this
as well, but they have just the one spear.

I incorporated this code into EvilHack because it will make the plans I
have for players setting their own traps that much better. I should be
able to now let the player do things like make an arrow trap out of
silver arrows, and have that trap actually fire silver arrows.