Crafting traps, part 2: fixes and enhancements.
Fair amount of work and testing went into this commit. We're still
at just being able to build an arrow trap - I won't expand the
selection until I get as many of the kinks worked out as I can. Here's
the breakdown:

* Player will receive more verbose feedback if their attempt to build a
trap fails (wrong component, not enough of the required component)
* A crafted trap will now carry over the components BUC, erosion state,
and material (example: using 10 silver arrows to build an arrow trap
will produce a 'silver arrow trap'; setting said trap, its ammo will be
silver arrows)
* Arrow traps will have the correct amount of arrows if made by the
player (initial creation, and when arrows are fired/trap is disarmed)

Found and fixed some bugs as well. Ammo weight was all over the place
for arrow/dart/bolt/spear traps. Ammo from a crafted arrow trap would be
seen as a tool rather a weapon (fired arrows as well as ammo from
disarming). Last bug: if the player was wielding a trap when they set
it, the game still thought it was wielded. In set_trap() the function
useup() was removed, and freeinv() was put in its place before
set_trap_ammo() was called. The useup() function would handle clearing
the owornmask, but freeinv() does not, so placing setnotworn() right
before freeinv() fixed that issue.

More to do, but we're getting there.