Allow object properties when crafting an arrow trap.
This commit allows arrows with an object property to have that property
carry over to its ammo when crafting an arrow trap. When the trap is
formally identified, it will show the object property its ammo has
(example: arrow trap of fire).

Regular arrow traps in the past have had arrows come with an object
property at random, but it was very rare, and it was always on a
per-fired-arrow basis, or when a trap was disarmed (mkobj randomness).
Now the player can purposely craft a trap using object properties. And
this of course will carry over to other ammo traps as they become
available to be built.