Fix: issue #181 (Convicts and killing hostile guards).
An issue was logged stating that the player was abusing their alignment
while playing as a Convict and killing the town watch that were hostile
to them.

First off - this scenario triggers that one bit in xkilled() where the
player's alignment can be adjusted, but no feedback is given. Secondly,
I agree that a Convict shouldn't get an alignment ding for killing
hostile watchmen/guards. Neither should Infidels for that matter (their
regular alignment record would be adversely affected, but not their
alignment abuse record. malign is funky).

This commit makes it so Convicts can kill the watch/guard without
abusing alignment if they're hostile, and Infidels can kill any monster
that is always peaceful without racking up a regular alignment record
penalty. It also provides 'you are guilty' feedback for those that did
abuse their alignment. Also fixed it so guilt feedback is given
if the player has to pay for shop damages. There's some code formatting
here as well.